LayoutTests/sputnik/Conformance/07_Lexical_Conventions/7.9_Automatic_Semicolon_Insertion/S7.9_A11_T4-expected.txt

CONSOLE MESSAGE: line 78: SyntaxError: Unexpected token 'else'
S7.9_A11_T4

PASS Expected parsing failure

TEST COMPLETE
